first look    v.1.0

+++ app/src/main/java/com/nimamoradi/NotePlus/databases/Items.java
@@ -22,7 +22,7 @@ public class Items {
         Title = title;
     }
 
-    public int getId() {
+    public long getId() {
         return id;
     }
